api: move validation methods to validation file
